Removed old login/register and switched to the new one
