+++ app/src/main/AndroidManifest.xml
@@ -18,16 +18,16 @@
             android:theme="@style/AppTheme"
             tools:ignore="GoogleAppIndexingWarning">
         <activity
-                android:name=".NewRegisterActivity"
-                android:label="@string/title_activity_new_register"
-                android:parentActivityName=".NewLoginActivity">
+                android:name=".RegisterActivity"
+                android:label="@string/title_activity_register"
+                android:parentActivityName=".LoginActivity">
             <meta-data
                     android:name="android.support.PARENT_ACTIVITY"
-                    android:value="com.lostanimals.tracks.NewLoginActivity"/>
+                    android:value="com.lostanimals.tracks.LoginActivity"/>
         </activity>
         <activity
-                android:name=".NewLoginActivity"
-                android:label="@string/title_activity_new_login">
+                android:name=".LoginActivity"
+                android:label="@string/title_activity_login">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
@@ -37,10 +37,6 @@
         <activity android:name=".FeedActivity">
         </activity>
         <activity android:name=".NewPostActivity">
-        </activity>
-        <activity android:name=".RegisterActivity">
-        </activity>
-        <activity android:name=".LoginActivity">
         </activity>
     </application>
