--- app/src/main/java/com/lostanimals/tracks/NewRegisterActivity.java
@@ -1,124 +1,0 @@
-package com.lostanimals.tracks;
-
-import android.annotation.TargetApi;
-import android.content.SharedPreferences;
-import android.os.Build;
-import android.os.Bundle;
-import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
-import android.text.TextUtils;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.widget.Button;
-import android.widget.EditText;
-import com.lostanimals.tracks.utils.EmailValidator;
-import com.lostanimals.tracks.utils.PreferencesUtility;
-import com.lostanimals.tracks.utils.ServerManager;
-
-public class NewRegisterActivity extends AppCompatActivity {
-
-    private ServerManager mServerTask = null;
-    private PreferencesUtility mPreferencesUtility;
-    // UI references.
-    private EditText mFirstNameView;
-    private EditText mLastNameView;
-    private EditText mEmailView;
-    private EditText mUsernameView;
-    private EditText mPasswordView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_new_register);
-        setupActionBar();
-
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        mPreferencesUtility = new PreferencesUtility(sharedPreferences);
-
-        mFirstNameView = findViewById(R.id.firstName);
-        mLastNameView = findViewById(R.id.lastName);
-        mEmailView = findViewById(R.id.email);
-        mUsernameView = findViewById(R.id.username);
-        mPasswordView = findViewById(R.id.password);
-
-        Button mRegisterBtn = findViewById(R.id.register_btn);
-        mRegisterBtn.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                attemptRegister();
-            }
-        });
-    }
-
-
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-    private void setupActionBar() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
-            // Show the Up button in the action bar.
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        }
-    }
-
-    private void attemptRegister() {
-
-        // Reset errors.
-        mFirstNameView.setError(null);
-        mLastNameView.setError(null);
-        mEmailView.setError(null);
-        mUsernameView.setError(null);
-        mPasswordView.setError(null);
-
-        // Store values at the time of the login attempt.
-        String firstName = mFirstNameView.getText().toString();
-        String lastName = mLastNameView.getText().toString();
-        String email = mEmailView.getText().toString();
-        String username = mUsernameView.getText().toString();
-        String password = mPasswordView.getText().toString();
-
-        boolean cancel = false;
-        View focusView = null;
-
-        // validate info
-        if (TextUtils.isEmpty(password)) {
-            mPasswordView.setError(getString(R.string.error_field_required));
-            focusView = mPasswordView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(username)) {
-            mUsernameView.setError(getString(R.string.error_field_required));
-            focusView = mUsernameView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(email)) {
-            mEmailView.setError(getString(R.string.error_field_required));
-            focusView = mEmailView;
-            cancel = true;
-        } else if (!(EmailValidator.isValidEmail(email))) {
-            mEmailView.setError(getString(R.string.email_incorrect));
-            focusView = mEmailView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(lastName)) {
-            mLastNameView.setError(getString(R.string.error_field_required));
-            focusView = mLastNameView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(firstName)) {
-            mFirstNameView.setError(getString(R.string.error_field_required));
-            focusView = mFirstNameView;
-            cancel = true;
-        }
-
-        String name = firstName + " " + lastName;
-
-        if (cancel) {
-            focusView.requestFocus();
-        } else {
-            mServerTask = new ServerManager(this);
-            mServerTask.setPreferencesUtility(mPreferencesUtility);
-            mServerTask.execute("register", name, email, username, password);
-        }
-    }
-
-}
-
